rm .travis.yml;
I haven't been getting much value out of the emscripten CI, and having
two CI systems is kind of awkward.  Also, travis doesn't let you change
the location of .travis.yml and it's cluttering up the root folder.  It
is possible to add this back in the future if there's a need for it,
preferably using appveyor or some other system.