gpu: ketchup experiment;
The idea is that we will greedily bump the GPU tick instead of
lazily letting it trail behind the CPU tick.  This should mean
that resources can get recycled (destroyed, readback, reused)
"as soon as" the GPU is done with them instead of 16 flushes later.

There is a function ketchup that catches the GPU tick up with the
current state of the fences.  Anyone that wants to query the GPU
tick can call ketchup beforehand to get a more up to date status.