gpu: don't ketchup unconditionally;
Checking fences every time a frame starts is actually taking a bit
of time.  Since we already waited on a fence, we do actually know
a little bit about where the GPU is at (it's at least completed the
frame for the fence we just waited on).  Instead of getting a more
fine-grained GPU tick by looking at all fences, we're just going to
set the GPU tick to at least tick[CPU] - COUNTOF(ticks) as a more
coarse method of ensuring forward progress that doesn't require
checking fences.  People that are desperately trying to recycle
resources are still encouraged to ketchup in case it allows stuff
to get recycled.