gpu: add ability to make a 1-view multiview pass;
Currently multiview is disabled if there's 0 or 1 views.

Shaders that use multiview seem to have issues when the subpass was
not created with a view mask (or have a zero view mask maybe).

To work around this, we need a way to create single-view multiview
passes.  To allow this, a view count of 1 will be created as multiview,
whereas a view count of zero will disable multiview.