Fix Pass:capsule when its length is zero;
Just draw a sphere.  The transform is rotated so the sphere segments
line up better, because spheres and capsules use different orientations
for their sphere parts.  Also the "degenerate" z axis is reconstructed
to be perpendicular to the x/z axes.  This doesn't seem like it will be
particularly fast, but hopefully people aren't drawing zero-length
capsules too often.  There might be an opportunity to shortcut the
rotation since it's 90 degrees and would just involve swapping columns.