glfw: filter resize events and cache window size;
GLFW reports window size as zero on Windows when the desktop window is
minimized.  This is by design.  Using zero width/height for window
textures isn't valid.  The fix is to ignore resize events where the
width or height is zero and also cache the last-valid window size so it
can be reported by os_window_get_size.  Sighs...