Fix morgue overflow;
The morgue is a fixed-size queue for GPU resources that are waiting to
be destroyed.  There's been an annoying issue with it for a while where
destroying too many objects at once will trigger a "Morgue overflow!"
error.  Even innocuous projects that create more than 1024 textures will
see this during a normal quit.

One way to solve this problem is to make the queue unbounded instead of
bounded.  However, this can hide problems and lead to more catastrophic
failure modes.

A better solution is to add "backpressure", where we avoid putting
things in the queue if it's full, or find some way to deal with them.
In this case it means finding a way to destroy stuff in the morgue when
it's full, to make space for more victims.

We weren't able to add backpressure reliably before, because command
buffers could have commands that reference the condemned resources.
This was mostly a problem for texture transfers -- if you create
thousands of textures in a loop, we'd have a giant command buffer with
commands to transfer pixels to the textures.  If these textures were
destroyed before submitting anything, the morgue would fill up, and we
wouldn't have any way to clear space because there was still a pending
command buffer that needs to act on the textures!

A simple change is to flush all pending transfers whenever a buffer or
texture is destroyed.  This lets us add backpressure to the morgue
because we can guarantee that there are no pending command buffers that
refer to an object in the morgue.

For backpressure, we try to destroy the oldest object in the morgue if
the GPU is done using it.  If that doesn't work, we'll wait on the fence
for its tick and destroy it.  This *should* always work, although in an
extreme case you could vkDeviceWaitIdle and clear out the entire morgue.

It should also be noted that in general command buffers need to be
flushed when destroying objects that they refer to.  However, for our
particular usage patterns, we only need to flush state.stream when a
buffer or texture is destroyed.  Pass objects already refcount their
buffers and textures and their commands are software command buffers, so
they don't require any special handling.  Other objects like shaders,
pipelines, descriptor set layouts, etc. all survive until shutdown, so
those don't impact anything either.